requirements.txt
asgiref==3.7.2
autopep8==2.0.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
cloudinary==1.36.0
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.8
django-allauth==0.57.0
django-crispy-forms==2.1
django-filter==23.5
django-summernote==0.8.20.0
djangorestframework==3.14.0
Flask==2.2.0
Flask-SQLAlchemy==2.5.1
google-auth==2.18.1
google-auth-oauthlib==1.0.0
greenlet==3.0.2
gspread==5.9.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
pillow==10.2.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.46
sqlparse==0.4.4
termcolor==2.3.0
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.0
wget==3.2
whitenoise==6.6.0
